+++ src/pages/Report/Report.tsx
@@ -1,15 +1,7 @@
-import { useState } from 'react'
 import { S } from '../Report/Report.style'
-import BasicButton from '../../components/common/BasicButton/BasicButton'
-import { theme } from '../../styles/theme'
+import ReportItem from '../../components/ReportItem/ReportItem'
 
 export default function Report() {
-  const [selectedButton, setSelectedButton] = useState<string | null>(null)
-
-  const handleButtonClick = (buttonText: string) => {
-    setSelectedButton(buttonText)
-  }
-
   // NOTICE: 임시 데이터
   const report = {
     title: '신고 받은 유저 리스트',
@@ -53,41 +45,7 @@ export default function Report() {
         </div>
       </S.TopTit>
       {reports.map((report, index) => (
-        <S.ContentBox key={index}>
-          <S.ListBox>
-            <p>{report.name}</p>
-            <p>{report.rating}</p>
-            <S.ButtonWrapper>
-              {report.buttons.map((button, buttonIndex) => (
-                <S.Button
-                  key={buttonIndex}
-                  onClick={() => handleButtonClick(button)}
-                  className={selectedButton === button ? 'selected' : ''}
-                >
-                  {button}
-                </S.Button>
-              ))}
-            </S.ButtonWrapper>
-          </S.ListBox>
-          <S.SubmitWrapper>
-            <BasicButton
-              type={'button'}
-              $bgColor={theme.colors.blue}
-              $width={'13rem'}
-              $fontcolor={theme.colors.white}
-            >
-              신고 내용
-            </BasicButton>
-            <BasicButton
-              type={'button'}
-              $bgColor={theme.colors.red}
-              $width={'13rem'}
-              $fontcolor={theme.colors.white}
-            >
-              블랙
-            </BasicButton>
-          </S.SubmitWrapper>
-        </S.ContentBox>
+        <ReportItem key={index} report={report} />
       ))}
     </S.Wrapper>
   )
